Split Command add Elasticsearch Connection

+++ src/Services/Products/ProductsServices.ts
@@ -1,9 +1,11 @@
 import {Product, IProduct} from "../../Models/Product";
 import {Services} from "../Services";
 import {MongoDB} from "../../Repository/MongoDB/MongoDB";
+import {JwtMiddleware} from "../../Middleware/JWT/JwtMiddleware";
+import Elasticsearch from "../../Repository/ElasticSearch/Elasticsearch";
 
 export class ProductsService extends Services {
-    constructor(private mongoDB: MongoDB) {
+    constructor(private mongoDB: MongoDB, private es: Elasticsearch, private jwt: JwtMiddleware) {
         super(mongoDB);
     }
 
@@ -53,8 +55,14 @@ export class ProductsService extends Services {
         }
         return product
     }
-    DeleteProduct = async (productId: String): Promise<any> => {
+    DeleteProduct = async (productId: String, token: String | undefined): Promise<any> => {
         let product: IProduct | null
+        const identity = this.jwt.GetIdentity(token)
+
+        if (identity.role != "Admin") {
+            return false
+        }
+
         try {
             await this.mongoDB.OpenConnection()
             product = await Product.findOneAndDelete({_id: productId})
